Handle payment sheet cancel event, by user
As the return type presentPaymentSheet method is Future<void>, waiting is done without any approval of payment been succeeded or failed (When the user closes or cancels the payment sheet, in spite, the code beneath was executing which was none other than showing payment success). 

At line 96, I've called retrievePaymentIntent to get the post-payment intent which includes the PaymentIntentsStatus and made a decision for both cases accordingly in a brief.