bugfix: android shapes fixed by using int values instead of double
'double' value type was creating issues with Android's rendering when tried to change payment sheet appearance. Using integer values is fixing this issue. That is of course a breaking change and needs to be discussed carefully.

Also added missing 'borderRadius' support for "PaymentSheetPrimaryButtonShape" class as this is already supported by both of the sdk's as well as in the stripe-react-native implementation.